pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "server-research-mcp"
version = "0.1.0"
description = "CrewAI MCP Research Server with Research Paper Parser"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.55.0",
    "anyio>=4.9.0",
    "asyncio>=3.4.3",
    "crewai>=0.95.0",
    "crewai-tools[mcp]>=0.47.1",
    "embedchain>=0.1.128",
    "gh>=0.0.4",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.15",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.2.14",
    "mcp>=1.9.4",
    "openai>=1.91.0",
    "psutil>=5.9.0",
    "pydantic>=2.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "re",
]

[tool.hatch.build.targets.wheel]
packages = ["src/server_research_mcp"]

[project.scripts]
run_crew = "server_research_mcp.main:run"
parse_paper = "server_research_mcp.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
markers = [
    "real_servers: marks tests that require real MCP servers (Node.js/npx)",
    "enhanced_mcp: marks tests that use enhanced MCP manager features",
    "performance: marks tests that measure performance metrics",
    "integration: marks integration tests that combine multiple components",
    "slow: marks tests that take a long time to run",
]
# Skip real server tests by default in CI
addopts = "-m 'not real_servers'"

[tool.mypy]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "interfaces.*"
strict = true
